Prevent integer overflow in computeDimensions. #12
Fix for CVE-2019-5086 and CVE-2019-5087

The code checks the sizes of width and height and stop execution, if it exceeds
maximal values.

+++ xcftools.h
@@ -121,7 +121,7 @@ FILE* openout(const char*);
 void closeout(FILE *,const char*);
 
 struct rect {
-  int t, b, l, r ;
+  int64_t t, b, l, r ;
 };
 
 #define isSubrect(A,B) \
